GoogleFrog (as we share a login right?)
Added slow weapon effect to an example copy of the Ion Cannon.
Also added hp bars as they display this effect.

It works as is so can just be added to any unit via  the config but it conflicts with MAX_SPEED in cob scripts. To work entirely well the scripts that use MAX_SPEED and are going to be affected by slow would need to be converted to lua. This is some of the battledroids and a few vehicles, not many.

git-svn-id: http://svn.imperialwinter.com/repo/trunk/swiw.sdd@3 c81d1ba1-7483-4aa6-992d-90018fdc1311